+++ frontend/src/Components/Footer.tsx
@@ -1,0 +1,10 @@
+function Footer(){
+
+    return(
+        <footer>
+            <p>&copy; {new Date().getFullYear()} Your website name</p>
+        </footer>
+    );
+}
+
+export default Footer
