+++ frontend/src/Components/Header.tsx
@@ -1,0 +1,10 @@
+function Header(){
+
+    return(
+        <header>
+            <h1>My WEBSITE</h1>
+        </header>
+    );
+}
+
+export default Header
